Improve test
- Add pytest-cov
- Refacto with `conf_test.py` for fixture
- Add pre-commit but not working yet (on my local workstation)
- Fix some IMPORTANT warning

+++ custom_components/frisquet_connect_unofficial/domains/site/site.py
@@ -66,7 +66,7 @@ class Site(ModelBase):
         return convert_from_epoch_to_datetime(int(self._date_derniere_remontee))
 
     @property
-    def external_temperature(self) -> int:
+    def external_temperature(self) -> float:
         return convert_api_temperature_to_float(self._external_temperature)
 
     @property
